refactor: Abstracting business logic from FavoriteProducts Context to useFavorite
